+++ jtransc-rt/src/java/lang/Object.java
@@ -64,8 +64,6 @@ import java.lang.reflect.Field;
 @JTranscAddFile(target = "as3", priority = -1, process = true, src = "as3/DivModResult.as", dst = "DivModResult.as")
 @JTranscAddFile(target = "as3", priority = -1, process = true, src = "as3/WrappedThrowable.as", dst = "WrappedThrowable.as")
 @JTranscAddFile(target = "as3", priority = -1, process = true, src = "as3/Main.xml", dst = "Main.xml")
-@JTranscAddLibraries(target = "cpp", value = "gc")
-@JTranscAddDefines(target = "cpp", value = "GC_NOT_DLL=1")
 @JTranscAddMembers(target = "d", value = "core.sync.mutex.Mutex __d_mutex = null;")
 public class Object {
 	@JTranscInvisible