+++ jtransc-rt/src/java/lang/Thread.java
@@ -27,9 +27,13 @@ import java.util.HashMap;
 import java.util.Map;
 
 @JTranscAddMembers(target = "d", value = "static {% CLASS java.lang.Thread %} _dCurrentThread; Thread thread;")
-//@JTranscAddHeader(target = "cpp", value = "#include <boost/thread.hpp>")
+@JTranscAddHeader(target = "cpp", cond = "USE_BOOST", value = {
+	"#include <boost/thread.hpp>",
+	"#include <boost/chrono.hpp>",
+})
 @JTranscAddIncludes(target = "cpp", value = "thread")
-@JTranscAddMembers(target = "cpp", value = "std::thread t_;")
+@JTranscAddMembers(target = "cpp", cond = "USE_BOOST", value = "boost::thread t_;")
+@JTranscAddMembers(target = "cpp", cond = "!USE_BOOST", value = "std::thread t_;")
 public class Thread implements Runnable {
 	public final static int MIN_PRIORITY = 1;
 	public final static int NORM_PRIORITY = 5;
@@ -64,12 +68,14 @@ public class Thread implements Runnable {
 	}
 
 	@JTranscMethodBody(target = "d", value = "Thread.sleep(dur!(\"msecs\")(p0));")
+	@JTranscMethodBody(target = "cpp", cond = "USE_BOOST", value = "boost::this_thread::sleep_for(boost::chrono::milliseconds(p0));")
 	@JTranscMethodBody(target = "cpp", value = "std::this_thread::sleep_for(std::chrono::milliseconds(p0));")
 	public static void sleep(long millis) throws InterruptedException {
 		JTranscSystem.sleep(millis);
 	}
 
 	@JTranscMethodBody(target = "d", value = "Thread.sleep(dur!(\"msecs\")(p0) + dur!(\"nsecs\")(p1));")
+	@JTranscMethodBody(target = "cpp", cond = "USE_BOOST", value = "boost::this_thread::sleep_for(boost::chrono::milliseconds(p0));")
 	@JTranscMethodBody(target = "cpp", value = "std::this_thread::sleep_for(std::chrono::milliseconds(p0));")
 	//FIXME
 	public static void sleep(long millis, int nanos) throws InterruptedException {
@@ -130,6 +136,7 @@ public class Thread implements Runnable {
 	}
 
 	@JTranscMethodBody(target = "d", value = "this.thread.start();")
+	@JTranscMethodBody(target = "cpp", cond = "USE_BOOST", value = "t_ = std::thread(&{% CLASS java.lang.Thread:runInternal %}::{% METHOD java.lang.Thread:runInternal:()V %}, this);")
 	@JTranscMethodBody(target = "cpp", value = "t_ = std::thread(&{% CLASS java.lang.Thread:runInternal %}::{% METHOD java.lang.Thread:runInternal:()V %}, this);")
 	public synchronized void start() {
 		JTranscThreading.impl.start(this);