+++ ChangeLog.md
@@ -4,6 +4,7 @@
 
 - Added `@JTranscAddIncludes` and `@JTranscAddDefines` annotations
 - Honor `@JTranscAddLibraries` and `@JTranscAddIncludes` annotations in C++ target
+- Added `@JTranscAddMembers.cond` to conditionally add members
 - Initial support of class `@JTranscNativeName` types in C++. Implemented Int32x4 SIMD to check. Works even without ENABLE_TYPING enabled.
 - Fxied compilation in windows (b2.exe not executing)